refactor: fix typo DataCatalogAutocomplete
